+++ GestorConversacion.py
@@ -1,112 +1,76 @@
-from config import bot
-import config
-from time import sleep
-import re
 import database.db as db
+from datetime import datetime
+from sqlalchemy import extract
+from models.Usuario import Usuario
 
-import logic
+def get_help():
+    response = (
+        "Estos son los comandos disponibles:\n"
+        "\n"
+        "*/start* - Inicio de la interacción con el bot\n\n"
+        "*/help* - Muestra este mensaje de ayuda\n\n"
+        "*/about* - Muestra detalles de esta aplicación y su equipo de desarrollo\n\n"
+        "*registrar paciente|rp {documento} {nombre completo}* - para registro de paciente\n\n"
+        "*registrar signos|rs {Presión arterial sistólica} {presión arterial diastólica} {frecuencia cardiaca, peso (kg)} {fecha (dd-mm-aaaa)} {hora (hh--mm AM/PM)}* - Para registro de signos vitales\n\n"
+        "*consultar signos|cs {Fecha inicial (dd-mm-aaaa)} {Fecha Final (dd-mm-aaaa)}* - para consultar sus signos registrados\n\n"
+        "*eliminar signos|es {número de la medición}* - eliminar medición, se recomienda consultar la medición para conocer su númeración\n\n"
+        "*consultar pacientes|cp {documento} {Fecha inicial (dd-mm-aaaa)} {Fecha Final (dd-mm-aaaa)}* - para realizar esta consulta debe estar habilitado como medico en nuestro sistema, permite consultar datos de pacientes \n\n"
+        "*ingresar observaciones|io {número de la medición} {observación asociada}* - permite a asociar una observación a una medición registrada por un paciente, funcionalidad solo disponible para medicos autorizados\n\n"
+        )
 
-from modules import  GestorConsultas
-from modules import  GestorMediciones
-from modules import  GestorObservaciones
-from modules import  GestorPacientes
+    return response
 
-#########################################################
-if __name__ == '__main__':
-    db.Base.metadata.create_all(db.engine)
-#########################################################
-
-# Comando de Inicio del bot
-@bot.message_handler(commands=['start'])
-def on_command_start(message):
-    bot.send_chat_action(message.chat.id, 'typing')
-
-    print (message.from_user)
-
-    bot.send_message (
-        message.chat.id,
-        logic.get_welcome_message(config.ASISTENTE_VIRTUAL,config.COMPANIA_SIGNOS),
-        parse_mode="Markdown")
+def get_paciente (documento):
+    usuario = db.session.query(Usuario).get(documento)
+    db.session.commit()
     
-    bot.send_message (
-        message.chat.id,
-        logic.get_help(),
-        parse_mode="Markdown")
+    if not usuario:
+        return None
     
-    #logic.register_account(message.from_user.id)
+    return usuario
+
+def set_paciente(documento, nombreCompleto): 
+    usuario = Usuario(documento, nombreCompleto)
+    
+    db.session.add(usuario)    
+    db.session.commit()
+
+    return True
 
 #########################################################
-
-@bot.message_handler(commands=['help'])
-def on_command_help(message):
-    bot.send_chat_action(message.chat.id, 'typing')
-    
-    bot.send_message(
-        message.chat.id,
-        logic.get_help(),
-        parse_mode="Markdown")
-
-@bot.message_handler(commands=['about'])
-def on_command_about(message):
-    pass
-
-
-@bot.message_handler(regexp=r"^(registrar signos|rs)$")
-def on_get_balance(message):
-    pass
-
-@bot.message_handler(regexp=r"^(registrar paciente|rp) ([0-9]*) ([a-zA-Z ]*)$")
-def on_get_balance(message):
-    bot.send_chat_action(message.chat.id, 'typing')
-
-    parts = re.match(r"^(registrar paciente|rp) ([0-9]*) ([a-zA-Z ]*)$",message.text)
-        
-    #print(parts[2])
-    #print(parts[3])
-
-    documento = int(parts[2])
-    nombreCompleto = parts[3]
-
-    usuario = logic.get_paciente(documento)
-    if usuario == None:
-        #Usuario no existente se procede al registro
-        logic.set_paciente(documento, nombreCompleto)
-        bot.reply_to(message, f"Paciente registrado.")
-    else:
-        bot.reply_to(message, f"Paciente ya registrado.")
-    
-
-
-@bot.message_handler(regexp=r"^(consultar signos|cs)$")
-def on_earn_money(message):
-    pass
-
-@bot.message_handler(regexp=r"^(eliminar signos|es)$")
-def on_spend_money(message):
-    pass
-
-@bot.message_handler(regexp=r"^(consultar pacientes|cp)$")
-def on_list_earnings(message):
-    pass
-
-@bot.message_handler(regexp=r"^(ingresar observaciones|io)$")
-def on_remove_record(message):
-    pass
-#############################################################################################
-# Mensaje por defecto que procesa los demás mensajes que coincidan 
-# con los comandos ingresados por el usuario
-
-@bot.message_handler(func=lambda message: True)
-def on_fallback(message):
-    bot.send_chat_action(message.chat.id, 'typing')
-    sleep(1) 
-
-    response = logic.get_fallback_message (message.text, message.from_user.first_name, config.COMPANIA_SIGNOS)
-    bot.reply_to(message, response, parse_mode="Markdown")
+# Lógica bienvenida e inicio al bot
+def get_welcome_message(ASISTENTE_VIRTUAL,COMPANIA_SIGNOS):
+    response = (
+            f"Hola, soy *{ASISTENTE_VIRTUAL}* el asistente virtual de *{COMPANIA_SIGNOS}* "
+            f"y quiero ayudarte con el registro y/o consulta de los signos vitales.\n\n"
+            "¡Recuerda que nuestra comunicación debe ser por texto !"
+        )
+    return response
 
 #########################################################
+# Lógica de registro del ID del usuario de telegram a la base de datos
+'''
+def register_account(user_id):
+    account = db.session.query(Account).get(user_id)
+    db.session.commit()
 
-if __name__ == '__main__':
-    bot.polling(timeout=1)
+    if account == None:
+        account = Account(user_id, 0)
+        db.session.add(account)
+        db.session.commit()
+        return True
+
+    return False
+'''
 #########################################################
-
+# Lógica del mensaje “fallback”
+def get_fallback_message (text,nombrePaciente,COMPANIA_SIGNOS):
+    response = (
+            f"\U0001F912 *{nombrePaciente}*, el comando o instrucción registrada no se identifica, "
+            f"revisa e intenta nuevamente.\n\n"
+            f"En *{COMPANIA_SIGNOS}* te acompañamos en el camino a tu bienestar, si requieres ayuda en cualquier "
+            f"momento puedes ejecutar el comando /help"
+   
+        )
+   
+    return response
